requirements.txt
alembic==1.0.7
aniso8601==4.1.0
asn1crypto==0.24.0
astroid==2.1.0
bcrypt==3.1.6
blinker==1.4
Brlapi==0.6.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
defer==1.0.6
elasticsearch==6.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
httplib2==0.9.2
idna==2.6
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.3.1
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
netifaces==0.10.4
olefile==0.45.1
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.2.2
pyOpenSSL==17.5.0
python-apt===1.6.3-elementary6-ubuntu5.0.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
schema==0.6.8
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
SQLAlchemy==1.2.17
typed-ast==1.2.0
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.24.1
virtualenv==16.3.0
virtualenv-clone==0.5.1
Werkzeug==0.14.1
wrapt==1.11.1
xkit==0.0.0
